requirements.txt
pyserial
bleak
